+++ core/src/main/java/org/apache/carbondata/core/keygenerator/KeyGenException.java
@@ -22,7 +22,7 @@ package org.apache.carbondata.core.keygenerator;
 /**
  * It can be thrown while generating the key.
  */
-public class KeyGenException extends Exception {
+public class KeyGenException extends RuntimeException {
 
   private static final long serialVersionUID = 3105132151795358241L;
