+++ processing/src/main/java/org/apache/carbondata/processing/newflow/compression/HeavyCompressorFactory.java
@@ -1,0 +1,19 @@
+package org.apache.carbondata.processing.newflow.compression;
+
+import org.apache.carbondata.core.carbon.metadata.blocklet.compressor.CompressionCodec;
+import org.apache.carbondata.processing.newflow.compression.impl.SnappyHeavyCompressorImpl;
+
+public class HeavyCompressorFactory {
+
+  private HeavyCompressorFactory() {
+
+  }
+  public static HeavyCompressor createHeavyCompressor(CompressionCodec compressionCodec) {
+    switch (compressionCodec) {
+      case SNAPPY :
+        return new SnappyHeavyCompressorImpl();
+      default:
+        return new SnappyHeavyCompressorImpl();
+    }
+  }
+}
