+++ processing/src/main/java/org/apache/carbondata/processing/newflow/encoding/ColumnData.java
@@ -1,0 +1,9 @@
+package org.apache.carbondata.processing.newflow.encoding;
+
+public interface ColumnData<E> {
+
+  void setColumnData(E input);
+
+  E getColumnData();
+
+}
