+++ processing/src/main/java/org/apache/carbondata/processing/newflow/encoding/ColumnEncoder.java
@@ -1,0 +1,11 @@
+package org.apache.carbondata.processing.newflow.encoding;
+
+import java.io.DataOutputStream;
+
+public interface ColumnEncoder {
+
+  void encode(ColumnData columnData, DataOutputStream stream);
+
+  byte[] encode(ColumnData columnData);
+
+}
