+++ processing/src/main/java/org/apache/carbondata/processing/newflow/encoding/FieldEncoder.java
@@ -1,0 +1,7 @@
+package org.apache.carbondata.processing.newflow.encoding;
+
+public interface FieldEncoder<E> {
+
+  E encode(Object[] data);
+
+}
