+++ processing/src/main/java/org/apache/carbondata/processing/newflow/encoding/RowEncoder.java
@@ -1,0 +1,7 @@
+package org.apache.carbondata.processing.newflow.encoding;
+
+public interface RowEncoder {
+
+  Object[] encode(Object[] row);
+
+}
