+++ processing/src/main/java/org/apache/carbondata/processing/newflow/encoding/impl/ComplexFieldEncoderImpl.java
@@ -1,0 +1,12 @@
+package org.apache.carbondata.processing.newflow.encoding.impl;
+
+import java.nio.ByteBuffer;
+
+import org.apache.carbondata.processing.newflow.encoding.FieldEncoder;
+
+public class ComplexFieldEncoderImpl implements FieldEncoder<ByteBuffer> {
+
+  @Override public ByteBuffer encode(Object[] data) {
+    return null;
+  }
+}
