+++ processing/src/main/java/org/apache/carbondata/processing/newflow/parser/GenericParser.java
@@ -1,0 +1,9 @@
+package org.apache.carbondata.processing.newflow.parser;
+
+public interface GenericParser<E> {
+
+  E parse(String data);
+
+  void addChildren(GenericParser parser);
+
+}
