+++ processing/src/main/java/org/apache/carbondata/processing/newflow/parser/impl/PrimitiveParserImpl.java
@@ -1,0 +1,14 @@
+package org.apache.carbondata.processing.newflow.parser.impl;
+
+import org.apache.carbondata.processing.newflow.parser.GenericParser;
+
+public class PrimitiveParserImpl implements GenericParser<Object> {
+
+  @Override public Object parse(String data) {
+    return data;
+  }
+
+  @Override public void addChildren(GenericParser parser) {
+    // No implementation
+  }
+}
