+++ processing/src/main/java/org/apache/carbondata/processing/newflow/steps/encodestep/EncoderProcessorStepImpl.java
@@ -1,0 +1,75 @@
+package org.apache.carbondata.processing.newflow.steps.encodestep;
+
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+
+import org.apache.carbondata.common.CarbonIterator;
+import org.apache.carbondata.core.cache.Cache;
+import org.apache.carbondata.core.cache.CacheProvider;
+import org.apache.carbondata.core.cache.CacheType;
+import org.apache.carbondata.core.cache.dictionary.Dictionary;
+import org.apache.carbondata.core.cache.dictionary.DictionaryColumnUniqueIdentifier;
+import org.apache.carbondata.processing.newflow.CarbonDataLoadConfiguration;
+
+import org.apache.carbondata.processing.newflow.DataField;
+import org.apache.carbondata.processing.newflow.DataLoadProcessorStep;
+import org.apache.carbondata.processing.newflow.encoding.FieldEncoder;
+import org.apache.carbondata.processing.newflow.encoding.impl.FieldEncoderFactory;
+
+/**
+ * Encode data with dictionary values and composed with bit/byte packed key.
+ * nondictionary values are packed as bytes, and complex types are also packed as bytes.
+ */
+public class EncoderProcessorStepImpl implements DataLoadProcessorStep {
+
+  private CarbonDataLoadConfiguration configuration;
+
+  private DataLoadProcessorStep child;
+
+  private FieldEncoder[] encoders;
+
+  @Override public DataField[] getOutput() {
+    return child.getOutput();
+  }
+
+  @Override
+  public void intialize(CarbonDataLoadConfiguration configuration, DataLoadProcessorStep child) {
+    this.configuration = configuration;
+    this.child = child;
+    CacheProvider cacheProvider = CacheProvider.getInstance();
+    Cache<DictionaryColumnUniqueIdentifier, Dictionary> cache =
+        cacheProvider.createCache(CacheType.REVERSE_DICTIONARY, configuration.getTableIdentifier().getStorePath());
+    List<FieldEncoder> fieldEncoders = new ArrayList<>();
+    DataField[] output = child.getOutput();
+    for (int i = 0; i < output.length; i++) {
+      FieldEncoder fieldEncoder = FieldEncoderFactory.getInstance()
+          .createFieldEncoder(output[i], cache,
+              configuration.getTableIdentifier().getCarbonTableIdentifier(), i);
+      fieldEncoders.add(fieldEncoder);
+    }
+    encoders = fieldEncoders.toArray(new FieldEncoder[fieldEncoders.size()]);
+    child.intialize(configuration, child);
+  }
+
+  @Override public Iterator<Object[]> execute() {
+    final Iterator<Object[]> iterator = child.execute();
+    return new CarbonIterator<Object[]>() {
+      @Override public boolean hasNext() {
+        return iterator.hasNext();
+      }
+
+      @Override public Object[] next() {
+        Object[] next = iterator.next();
+        for (int i = 0; i < encoders.length; i++) {
+          encoders[i].encode(next);
+        }
+        return next;
+      }
+    };
+  }
+
+  @Override public void close() {
+
+  }
+}
