+++ processing/src/main/java/org/apache/carbondata/processing/newflow/steps/input/InputProcessorStepImpl.java
@@ -1,0 +1,92 @@
+package org.apache.carbondata.processing.newflow.steps.input;
+
+import java.io.IOException;
+import java.util.Iterator;
+
+import org.apache.carbondata.common.CarbonIterator;
+import org.apache.carbondata.processing.newflow.CarbonArrayWritable;
+import org.apache.carbondata.processing.newflow.CarbonDataLoadConfiguration;
+
+import org.apache.hadoop.mapred.RecordReader;
+
+import org.apache.carbondata.processing.newflow.DataField;
+import org.apache.carbondata.processing.newflow.DataLoadProcessorStep;
+import org.apache.carbondata.processing.newflow.parser.CarbonParserFactory;
+import org.apache.carbondata.processing.newflow.parser.GenericParser;
+
+/**
+ * It is wrapper class around input format
+ */
+public class InputProcessorStepImpl implements DataLoadProcessorStep {
+
+  private RecordReader<Void, CarbonArrayWritable> recordReader;
+
+  private CarbonDataLoadConfiguration configuration;
+
+  private GenericParser[] genericParsers;
+
+  @Override public DataField[] getOutput() {
+    DataField[] fields = configuration.getDataFields();
+    String[] header = configuration.getHeader();
+    DataField[] output = new DataField[fields.length];
+    int k = 0;
+    for (int i = 0; i < header.length; i++) {
+      for (int j = 0; j < fields.length; j++) {
+        if (header[j].equalsIgnoreCase(fields[j].getColumn().getColName())) {
+          output[k++] = fields[j];
+          break;
+        }
+      }
+    }
+    return output;
+  }
+
+  @Override
+  public void intialize(CarbonDataLoadConfiguration configuration, DataLoadProcessorStep child) {
+    this.recordReader = configuration.getRecordReader();
+    this.configuration = configuration;
+    DataField[] output = getOutput();
+    genericParsers = new GenericParser[output.length];
+    for (int i = 0; i < genericParsers.length; i++) {
+      genericParsers[i] = CarbonParserFactory
+          .createParser(output[i].getColumn(), configuration.getComplexDelimiters());
+    }
+
+    if (child != null) {
+      child.intialize(configuration, child);
+    }
+  }
+
+  @Override public Iterator<Object[]> execute() {
+
+    return new CarbonIterator<Object[]>() {
+      CarbonArrayWritable data = new CarbonArrayWritable();
+
+      @Override public boolean hasNext() {
+        try {
+          return recordReader.next(null, data);
+        } catch (IOException e) {
+          e.printStackTrace();
+        }
+        return false;
+      }
+
+      @Override public Object[] next() {
+
+        Object[] row = data.get();
+        for (int i = 0; i < row.length; i++) {
+          row[i] = genericParsers[i].parse(row[i].toString());
+        }
+        return row;
+      }
+    };
+  }
+
+  @Override public void close() {
+    try {
+      this.recordReader.close();
+    } catch (IOException e) {
+      e.printStackTrace();
+    }
+  }
+}
