--- processing/src/main/java/org/carbondata/processing/newflow/complexobjects/ArrayObject.java
@@ -1,18 +1,0 @@
-package org.carbondata.processing.newflow.complexobjects;
-
-public class ArrayObject {
-
-  private Object[] data;
-
-  public ArrayObject(Object[] data) {
-    this.data = data;
-  }
-
-  public Object[] getData() {
-    return data;
-  }
-
-  public void setData(Object[] data) {
-    this.data = data;
-  }
-}
