--- processing/src/main/java/org/carbondata/processing/newflow/encoding/ColumnData.java
@@ -1,9 +1,0 @@
-package org.carbondata.processing.newflow.encoding;
-
-public interface ColumnData<E> {
-
-  void setColumnData(E input);
-
-  E getColumnData();
-
-}
