--- processing/src/main/java/org/carbondata/processing/newflow/encoding/ColumnEncoder.java
@@ -1,11 +1,0 @@
-package org.carbondata.processing.newflow.encoding;
-
-import java.io.DataOutputStream;
-
-public interface ColumnEncoder {
-
-  void encode(ColumnData columnData, DataOutputStream stream);
-
-  byte[] encode(ColumnData columnData);
-
-}
