--- processing/src/main/java/org/carbondata/processing/newflow/encoding/FieldEncoder.java
@@ -1,7 +1,0 @@
-package org.carbondata.processing.newflow.encoding;
-
-public interface FieldEncoder {
-
-  void encode(Object[] data);
-
-}
