--- processing/src/main/java/org/carbondata/processing/newflow/encoding/RowEncoder.java
@@ -1,7 +1,0 @@
-package org.carbondata.processing.newflow.encoding;
-
-public interface RowEncoder {
-
-  Object[] encode(Object[] row);
-
-}
