--- processing/src/main/java/org/carbondata/processing/newflow/parser/GenericParser.java
@@ -1,9 +1,0 @@
-package org.carbondata.processing.newflow.parser;
-
-public interface GenericParser<E> {
-
-  E parse(String data);
-
-  void addChildren(GenericParser parser);
-
-}
